--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -1,0 +1,33 @@
+# Include important files in the package distribution
+include README.md
+include LICENSE
+include CHANGELOG.md
+include requirements.txt
+include requirements-gpu.txt
+
+# Include configuration files
+recursive-include config *.py *.yml *.yaml *.json
+
+# Include source code
+recursive-include src *.py
+
+# Include Docker and deployment files
+recursive-include docker *.sh *.yml *.yaml Dockerfile*
+recursive-include deployment *.sh *.yml *.yaml
+recursive-include aws *.yml *.yaml *.sh
+
+# Include documentation
+recursive-include docs *.md *.rst *.txt
+
+# Exclude test files from distribution
+prune tests
+prune *.egg-info
+
+# Exclude temporary and cache files
+global-exclude __pycache__
+global-exclude *.py[co]
+global-exclude .DS_Store
+global-exclude *.so
+global-exclude .git*
+global-exclude *.tmp
+global-exclude *.log