--- a/pytest.ini
+++ b/pytest.ini
@@ -1,0 +1,20 @@
+[tool:pytest]
+minversion = 7.0
+addopts = 
+    -ra 
+    -q 
+    --strict-markers
+    --disable-warnings
+    --tb=short
+testpaths = tests
+markers =
+    unit: Unit tests for individual components
+    integration: Integration tests for component interactions
+    performance: Performance and benchmark tests
+    slow: Tests that take longer to run
+    gpu: Tests that require GPU hardware
+    aws: Tests that require AWS services or mocking
+filterwarnings =
+    ignore::UserWarning
+    ignore::DeprecationWarning
+    ignore::FutureWarning