--- a/src/pipeline/__init__.py
+++ b/src/pipeline/__init__.py
@@ -1,0 +1,3 @@
+"""
+Pipeline Management - Workflow Orchestration and Resource Management
+"""