… CHANGELOG.md
… CLAUDE.md
… README.md
… aws/cloudformation.yml
… deployment/aws-deploy.sh
… docker-compose.aws.yml
… docker/.env.template
… docker/Dockerfile.gpu
… docker/README.md
… docker/docker-entrypoint.sh
… docker/healthcheck.sh
… main.py
… output/friends_analysis_20250830_171903.json
… output/friends_analysis_20250830_172706.json
… output/friends_analysis_20250830_174820.json
… output/optimized_friends_test.json
… pyproject.toml
… requirements-gpu.txt
… requirements.txt
… run_tests.py
… scripts/build_package.sh
… setup.py
… src/__init__.py
… src/api.py
… src/cli.py
… src/models/__init__.py
… src/models/emotion_analyzer.py
… src/models/model_manager.py
… src/models/output_models.py
… src/models/speaker_diarizer.py